+++ scripts/commands/updateApiDocs.ts
@@ -31,8 +31,7 @@ import { specialCaseResults } from "../lib/api/specialCaseResults";
 import addFrontMatter from "../lib/api/addFrontMatter";
 import { dedupeHtmlIdsFromResults } from "../lib/api/dedupeHtmlIds";
 import { removePrefix, removeSuffix } from "../lib/stringUtils";
-import { Link } from "../lib/sharedTypes";
-import { Pkg, PkgInfo, getPkgRoot } from "../lib/api/Pkg";
+import { Link, Pkg, PkgInfo, getPkgRoot } from "../lib/api/Pkg";
 import { zxMain } from "../lib/zx";
 import { pathExists, getRoot, rmFilesInFolder } from "../lib/fs";
 import { downloadCIArtifact } from "../lib/api/downloadCIArtifacts";
